Generic version download script for JDK
Converted the download script to be generic, so not hard-coded to a specific
binary version.

The following Variables have been created

JDK_CURRENT_VERSION
Used to set the current version number as a string.  This is the only
variable that needs to be updated when a new binary is available.

The name and download path for the binary file are combined in the
JDK_TAR_FILE_NAME variable with the follwing variables and the current
version variable

JDK_NAME_GENERIC="jigsaw-jdk-9-ea+"
JDK_NAME_LINUX="_linux-x64_bin.tar.gz"
JDK_NAME_OSX="_osx-x64_bin.tar.gz"

The wget command has been updated to include the JDK_CURRENT_VERSION
variable in the path for the java.net website.